Moved AlertDialogFragment and ColorWheel java classes to ui package as well as eliminate redundant returns.
